test: split test-serve
Avoids timeouts on Windows debug bots.
