+++ tests/test-serve-custom-port.js
@@ -1,0 +1,39 @@
+import assert from 'tjs:assert';
+
+import { spawnServe } from './helpers/serve-spawn.js';
+
+
+async function testServeCustomPort() {
+    const { proc, port } = await spawnServe('serve-simple.js', [ '--port', '0' ]);
+
+    try {
+        assert.ok(port > 0, 'port was assigned');
+
+        const resp = await fetch(`http://127.0.0.1:${port}/`);
+        assert.eq(resp.status, 200, 'status is 200');
+
+        const text = await resp.text();
+        assert.eq(text, 'hello from serve', 'body matches');
+    } finally {
+        proc.kill('SIGTERM');
+        await proc.wait();
+    }
+}
+
+async function testServeCustomPortShort() {
+    const { proc, port } = await spawnServe('serve-simple.js', [ '-p', '0' ]);
+
+    try {
+        assert.ok(port > 0, 'port was assigned');
+
+        const resp = await fetch(`http://127.0.0.1:${port}/`);
+        const text = await resp.text();
+        assert.eq(text, 'hello from serve', 'body matches with -p flag');
+    } finally {
+        proc.kill('SIGTERM');
+        await proc.wait();
+    }
+}
+
+await testServeCustomPort();
+await testServeCustomPortShort();
